session and suggestions controllers created. added notes for spotifyitem model
