+++ app/controllers/session_controller.rb
@@ -1,0 +1,2 @@
+class SessionController < ActionController::Base
+end
