+++ app/controllers/suggestions_controller.rb
@@ -1,0 +1,2 @@
+class suggestionsController < ActionController::Base
+end
